# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for fetching and managing stock ratings data",
        "title": "Smart Stock Recommender API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
        "/stocks": {
            "post": {
                "description": "Retrieves stock data from external API for a specific page and stores in database. Returns the raw API response with stock items and next page token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "Fetch stocks by page number",
                "parameters": [
                    {
                        "description": "Request body with page number (integer, required)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully fetched stock data from external API",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid JSON format, missing page field, or invalid page number",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error occurred",
                        "schema": {
                            "$ref": "#/definitions/models.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/stocks/bulk": {
            "post": {
                "description": "Clears existing database data, then fetches stock data from external API for a range of pages using parallel processing. Returns summary statistics of the operation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "Fetch stocks in bulk for page range with parallel processing",
                "parameters": [
                    {
                        "description": "Request body with start_page and end_page (integers, both required, max range 1,000,000)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BulkPageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully processed bulk stock data fetch with parallel processing",
                        "schema": {
                            "$ref": "#/definitions/models.BulkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid JSON, negative pages, start \u003e end, or range too large",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error occurred",
                        "schema": {
                            "$ref": "#/definitions/models.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/stocks/list": {
            "post": {
                "description": "Retrieves stored stock ratings with pagination support, ordered by creation date (newest first). Returns both data and pagination metadata.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stocks"
                ],
                "summary": "Get paginated stock ratings from database",
                "parameters": [
                    {
                        "description": "Request body with page_number (integer, min 1) and page_length (integer, 1-1000)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PaginationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved paginated stock ratings with metadata",
                        "schema": {
                            "$ref": "#/definitions/models.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid JSON, page_number \u003c= 0, or page_length not between 1-1000",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error occurred",
                        "schema": {
                            "$ref": "#/definitions/models.GenericErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ApiResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StockRatings"
                    }
                },
                "next_page": {
                    "type": "string"
                }
            }
        },
...
        "models.BulkPageRequest": {
            "type": "object",
            "required": [
                "end_page",
                "start_page"
            ],
            "properties": {
                "end_page": {
                    "type": "integer",
                    "example": 100
                },
                "start_page": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.BulkResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Successfully fetched and stored stock data"
                },
                "pages_fetched": {
                    "type": "string",
                    "example": "1-1000"
                },
                "stocks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StockRatings"
                    }
                },
                "total_stocks": {
                    "type": "integer",
                    "example": 7860
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid JSON format in request body"
                }
            }
        },
        "models.GenericErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Internal server error occurred"
                }
            }
        },
...
        "models.PageRequest": {
            "type": "object",
            "required": [
                "page"
            ],
            "properties": {
                "page": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StockRatings"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/models.PaginationMeta"
                }
            }
        },
        "models.PaginationMeta": {
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean",
                    "example": true
                },
                "has_previous": {
                    "type": "boolean",
                    "example": false
                },
                "page_length": {
                    "type": "integer",
                    "example": 20
                },
                "page_number": {
                    "type": "integer",
                    "example": 1
                },
                "total_pages": {
                    "type": "integer",
                    "example": 126
                },
                "total_records": {
                    "type": "integer",
                    "example": 2520
                }
            }
        },
        "models.PaginationRequest": {
            "type": "object",
            "required": [
                "page_length",
                "page_number"
            ],
            "properties": {
                "page_length": {
                    "type": "integer",
                    "example": 20
                },
                "page_number": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.StockRatings": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "target raised by"
                },
                "brokerage": {
                    "type": "string",
                    "example": "Goldman Sachs"
                },
                "company": {
                    "type": "string",
                    "example": "Apple Inc."
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-01-15T10:35:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "rating_from": {
                    "type": "string",
                    "example": "Buy"
                },
                "rating_to": {
                    "type": "string",
                    "example": "Strong Buy"
                },
                "target_from": {
                    "type": "string",
                    "example": "$150.00"
                },
                "target_to": {
                    "type": "string",
                    "example": "$180.00"
                },
                "ticker": {
                    "type": "string",
                    "example": "AAPL"
                },
                "time": {
                    "type": "string",
                    "example": "2025-01-15T10:30:00Z"
                }
            }
        },
...
    }
}